Layout uses global context for getting application theme, toggle it and save search term input
